app/requirements.txt
asgiref==3.7.2
Django==4.2.6
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
protobuf==4.21.12
sqlparse==0.4.4
tzdata==2023.3
python-docx==0.8.11
python-dotenv==1.0.0
python-pptx==0.6.22
openai==0.27.4
PyPDF2==3.0.1
